remove submodule, move pages to prop passed to header
