+++ .env
@@ -1,1 +1,2 @@
 REACT_APP_API_HOST=https://api.opendota.com
+BROWSER=false
