--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "src/common"]
-	path = src/common
-	url = https://github.com/odota/components
