+++ src/components/Header/Header.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
 import ActionSearch from 'material-ui/svg-icons/action/search';
 import { Toolbar, ToolbarGroup } from 'material-ui/Toolbar';
 import IconButton from 'material-ui/IconButton';
@@ -89,21 +88,8 @@ class Header extends React.Component {
 
   render() {
     const {
-      location, small, user, strings,
+      location, small, user, strings, navbarPages
     } = this.props;
-    const navbarPages = [
-      <Link key="header_explorer" to="/explorer">{strings.header_explorer}</Link>,
-      <Link key="header_meta" to="/meta">{strings.header_meta}</Link>,
-      <Link key="header_matches" to="/matches">{strings.header_matches}</Link>,
-      <Link key="header_teams" to="/teams">{strings.header_teams}</Link>,
-      <Link key="header_heroes" to="/heroes">{strings.header_heroes}</Link>,
-      <Link key="header_distributions" to="/distributions">{strings.header_distributions}</Link>,
-      <Link key="header_records" to="/records">{strings.header_records}</Link>,
-      <Link key="header_scenarios" to="/scenarios">{strings.header_scenarios}</Link>,
-      <Link key="header_api" to="/api-keys">{strings.header_api}</Link>,
-      // <Link key="header_predictions" to="/predictions">TI Predictions</Link>,
-      // <Link key="header_assistant" to="/assistant">Assistant</Link>,
-    ];
 
     const burgerItems = [
       <AccountWidget key={0} />,
